chore: Update site theme to "Sunset Glow"
The code changes in the `.deco/blocks/site.json` file update the site theme to "Sunset Glow" by changing the value of the `__resolveType` property. This improves the visual appearance of the site.

Note: This suggested commit message follows the established convention of starting with a type (chore) followed by a concise and descriptive summary of the changes.